requirements.txt
attrdict==2.0.1
certifi==2019.3.9
chardet==3.0.4
cytoolz==0.9.0.1
dj-database-url==0.5.0
Django==2.2
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.1
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.5.1
gunicorn==19.9.0
hexbytes==0.1.0
idna==2.8
lru-dict==1.1.6
parsimonious==0.8.1
psycopg2==2.8.2
pycryptodome==3.8.1
python-dotenv==0.10.1
pytz==2019.1
requests==2.21.0
rlp==1.1.0
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
urllib3==1.24.2
web3==4.9.2
websockets==6.0
whitenoise==4.1.2
